rename download_graph / dbt_graph, update docs
