+++ docs/content-crag/integrations/dbt.mdx
@@ -21,7 +21,7 @@ The resources that dagster_dbt provides are
 
 Both create resources that derive from the same base class, <PyObject module='dagster_dbt' object='DbtResource'/>.
 
-This library also provides pre-built solids that can be integrated into your pipelines, but we recommend using the resources for most use cases, as they provide much more flexibility in how you interact with dbt. To view example pipelines that take advantage of this dbt integration, check out either the [Hacker News example repo](https://github.com/dagster-io/dagster/blob/master/examples/hacker_news/hacker_news/pipelines/dbt_pipeline.py) or the [dbt example repo](https://github.com/dagster-io/dagster/tree/master/examples/dbt_example).
+This library also provides pre-built solids that can be integrated into your pipelines, but we recommend using the resources for most use cases, as they provide much more flexibility in how you interact with dbt. To view example pipelines that take advantage of this dbt integration, check out either the [Hacker News example repo](https://github.com/dagster-io/dagster/blob/master/examples/hacker_news/hacker_news/jobs/dbt_metrics.py) or the [dbt example repo](https://github.com/dagster-io/dagster/tree/master/examples/dbt_example).
 
 ## Using the dbt CLI resource in a Dagster pipeline
 
@@ -398,11 +398,6 @@ If you have questions on using dbt with Dagster, we'd love to hear from you:
 
 <p align="center">
   <a href="https://dagster-slackin.herokuapp.com/" target="_blank">
-    <Image
-      alt="join-us-on-slack"
-      src="/assets/join-us-on-slack.png"
-      width="160"
-      height="40"
-    />
+    <Image alt="join-us-on-slack" src="/assets/join-us-on-slack.png" width="160" height="40" />
   </a>
 </p>
