+++ docs/content/integrations/dbt.mdx
@@ -26,7 +26,7 @@ The resources that dagster_dbt provides are
 
 Both create resources that derive from the same base class, <PyObject module='dagster_dbt' object='DbtResource'/>.
 
-This library also provides pre-built solids that can be integrated into your pipelines, but we recommend using the resources for most use cases, as they provide much more flexibility in how you interact with dbt. To view example pipelines that take advantage of this dbt integration, check out either the [Hacker News example repo](https://github.com/dagster-io/dagster/blob/master/examples/hacker_news/hacker_news/pipelines/dbt_pipeline.py) or the [dbt example repo](https://github.com/dagster-io/dagster/tree/master/examples/dbt_example).
+This library also provides pre-built solids that can be integrated into your pipelines, but we recommend using the resources for most use cases, as they provide much more flexibility in how you interact with dbt. To view example pipelines that take advantage of this dbt integration, check out either the [Hacker News example repo](https://github.com/dagster-io/dagster/blob/master/examples/hacker_news/hacker_news/jobs/dbt_metrics.py) or the [dbt example repo](https://github.com/dagster-io/dagster/tree/master/examples/dbt_example).
 
 ## Using the dbt CLI resource in a Dagster pipeline
