+++ examples/hacker_news/hacker_news_tests/test_jobs/test_download_job.py
@@ -1,7 +1,7 @@
 import tempfile
 
 from dagster import ResourceDefinition, fs_io_manager
-from hacker_news.jobs.download_job import configured_pyspark, download_graph
+from hacker_news.jobs.hacker_news_api_download import configured_pyspark, download_graph
 from hacker_news.resources.hn_resource import hn_snapshot_client
 from hacker_news.resources.parquet_io_manager import partitioned_parquet_io_manager
